restructure examples v2 (#954)
- No more -lazy symlinks
- No more -mpi in name (the only non-MPI example, wave.py has been renamed to wave-nompi.py).
- Test accuracy of all examples, comparing eager/lazy/numpy, as part of run_examples.sh
