+++ .github/workflows/ci.yaml
@@ -117,20 +117,12 @@ jobs:
           run: |
             . .ci-support/install.sh
 
-        - name: Test lazy/numpy accuracy to eager
-          run: |
-            MINIFORGE_INSTALL_DIR=.miniforge3
-            . "$MINIFORGE_INSTALL_DIR/bin/activate" testing
-            mamba install vtk  # needed for the accuracy comparison
-            export XDG_CACHE_HOME=/tmp
-            [[ $(hostname) == "porter" ]] && export PYOPENCL_TEST="port:nv" && unset XDG_CACHE_HOME
-            # && export POCL_DEBUG=cuda
-            scripts/run-integrated-tests.sh --lazy-accuracy
         - name: Run examples
           run: |
             MINIFORGE_INSTALL_DIR=.miniforge3
             . "$MINIFORGE_INSTALL_DIR/bin/activate" testing
             export XDG_CACHE_HOME=/tmp
+            mamba install vtk  # needed for the accuracy comparison
             [[ $(hostname) == "porter" ]] && export PYOPENCL_TEST="port:nv" && unset XDG_CACHE_HOME
             # && export POCL_DEBUG=cuda
             scripts/run-integrated-tests.sh --examples
@@ -188,7 +180,7 @@ jobs:
             cd ..
             source emirge/config/activate_env.sh
             cd mirgecom/examples
-            python -m mpi4py ./pulse-mpi.py
+            python -m mpi4py ./pulse.py
 
     production:
         name: Production testing
