--- examples/ablation-workshop-mpi-lazy.py
@@ -1,1 +1,0 @@
-ablation-workshop-mpi.py
