--- examples/autoignition-mpi-lazy.py
@@ -1,1 +1,0 @@
-autoignition-mpi.py
