--- examples/doublemach-mpi-lazy.py
@@ -1,1 +1,0 @@
-doublemach-mpi.py
