--- examples/heat-source-mpi-lazy.py
@@ -1,1 +1,0 @@
-heat-source-mpi.py
