+++ examples/heat-source.py
