--- examples/hotplate-mpi-lazy.py
@@ -1,1 +1,0 @@
-hotplate-mpi.py
