+++ examples/hotplate.py
@@ -1,4 +1,4 @@
-"""Demonstrate simple mass lump advection."""
+"""Demonstrate a fluid between two hot plates in 2d."""
 
 __copyright__ = """
 Copyright (C) 2020 University of Illinois Board of Trustees
@@ -27,25 +27,30 @@ import logging
 import numpy as np
 from functools import partial
 
-from grudge.dof_desc import DISCR_TAG_QUAD
 from meshmode.mesh import BTAG_ALL, BTAG_NONE  # noqa
+
 from grudge.shortcuts import make_visualizer
+from grudge.dof_desc import BoundaryDomainTag
 
 from mirgecom.discretization import create_discretization_collection
-from mirgecom.euler import euler_operator
+from mirgecom.fluid import make_conserved
+from mirgecom.navierstokes import ns_operator
 from mirgecom.simutil import (
     get_sim_timestep,
-    generate_and_distribute_mesh
+    get_box_mesh
 )
+
 from mirgecom.io import make_init_message
 from mirgecom.mpi import mpi_entry_point
-
 from mirgecom.integrators import rk4_step
 from mirgecom.steppers import advance_state
-from mirgecom.boundary import PrescribedFluidBoundary
-from mirgecom.initializers import Lump, Uniform  # noqa
+from mirgecom.boundary import (
+    PrescribedFluidBoundary,
+    IsothermalWallBoundary
+)
+from mirgecom.transport import SimpleTransport
 from mirgecom.eos import IdealSingleGas
-
+from mirgecom.gas_model import GasModel, make_fluid_state
 from logpyle import IntervalTimer, set_dt
 from mirgecom.euler import extract_vars_for_logging, units_for_logging
 from mirgecom.logging_quantities import (
@@ -55,6 +60,7 @@ from mirgecom.logging_quantities import (
     logmgr_add_device_memory_usage,
     set_sim_state
 )
+
 
 logger = logging.getLogger(__name__)
 
@@ -66,10 +72,9 @@ class MyRuntimeError(RuntimeError):
 
 
 @mpi_entry_point
-def main(actx_class, use_esdg=False,
-         use_overintegration=False,
-         use_leap=False, casename=None, rst_filename=None):
-    """Drive example."""
+def main(actx_class, use_esdg=False, use_leap=False, casename=None,
+         use_overintegration=False, rst_filename=None):
+    """Drive the example."""
     if casename is None:
         casename = "mirgecom"
 
@@ -90,26 +95,28 @@ def main(actx_class, use_esdg=False,
     use_profiling = actx_class_is_profiling(actx_class)
 
     # timestepping control
-    if use_leap:
-        from leap.rk import RK4MethodBuilder
-        timestepper = RK4MethodBuilder("state")
-    else:
-        timestepper = rk4_step
-    t_final = .005
-    current_cfl = 1.0
-    current_dt = .001
+    timestepper = rk4_step
+    t_final = 2e-7
+    current_cfl = .1
+    current_dt = 1e-8
     current_t = 0
-    current_step = 0
     constant_cfl = False
+    current_step = 0
 
     # some i/o frequencies
     nstatus = 1
+    nviz = 100
+    nrestart = 100
     nhealth = 1
-    nrestart = 10
-    nviz = 1
 
+    # some geometry setup
     dim = 2
-
+    if dim != 2:
+        raise ValueError("This example must be run with dim = 2.")
+    left_boundary_location = 0
+    right_boundary_location = 0.1
+    bottom_boundary_location = 0
+    top_boundary_location = .02
     rst_path = "restart_data/"
     rst_pattern = (
         rst_path + "{cname}-{step:04d}-{rank:04d}.pkl"
@@ -122,28 +129,29 @@ def main(actx_class, use_esdg=False,
         local_mesh = restart_data["local_mesh"]
         local_nelements = local_mesh.nelements
         global_nelements = restart_data["global_nelements"]
-        assert restart_data["num_parts"] == nparts
+        assert restart_data["nparts"] == nparts
     else:  # generate the grid from scratch
-        box_ll = -5.0
-        box_ur = 5.0
-        nel_1d = 16
-        from meshmode.mesh.generation import generate_regular_rect_mesh
-        generate_mesh = partial(generate_regular_rect_mesh, a=(box_ll,)*dim,
-                                b=(box_ur,) * dim, nelements_per_axis=(nel_1d,)*dim)
+        n_refine = 2
+        nels_x = 5 * n_refine
+        nels_y = 3 * n_refine
+        nels_axis = (nels_x, nels_y)
+        box_ll = (left_boundary_location, bottom_boundary_location)
+        box_ur = (right_boundary_location, top_boundary_location)
+        generate_mesh = partial(get_box_mesh, 2, a=box_ll, b=box_ur, n=nels_axis)
+        from mirgecom.simutil import generate_and_distribute_mesh
         local_mesh, global_nelements = generate_and_distribute_mesh(comm,
                                                                     generate_mesh)
         local_nelements = local_mesh.nelements
 
-    order = 3
+    order = 1
     dcoll = create_discretization_collection(actx, local_mesh, order=order)
     nodes = actx.thaw(dcoll.nodes())
 
+    from grudge.dof_desc import DISCR_TAG_QUAD
     if use_overintegration:
         quadrature_tag = DISCR_TAG_QUAD
     else:
         quadrature_tag = None
-
-    vis_timer = None
 
     if logmgr:
         logmgr_add_cl_device_info(logmgr, queue)
@@ -151,30 +159,49 @@ def main(actx_class, use_esdg=False,
         logmgr_add_many_discretization_quantities(logmgr, dcoll, dim,
                              extract_vars_for_logging, units_for_logging)
 
-        vis_timer = IntervalTimer("t_vis", "Time spent visualizing")
-        logmgr.add_quantity(vis_timer)
-
         logmgr.add_watches([
             ("step.max", "step = {value}, "),
             ("t_sim.max", "sim time: {value:1.6e} s\n"),
             ("min_pressure", "------- P (min, max) (Pa) = ({value:1.9e}, "),
             ("max_pressure",    "{value:1.9e})\n"),
+            ("min_temperature", "------- T (min, max) (K) = ({value:1.9e}, "),
+            ("max_temperature",    "{value:1.9e})\n"),
             ("t_step.max", "------- step walltime: {value:6g} s, "),
             ("t_log.max", "log walltime: {value:6g} s")
         ])
 
-    # soln setup, init
-    eos = IdealSingleGas()
-    vel = np.zeros(shape=(dim,))
-    orig = np.zeros(shape=(dim,))
-    vel[:dim] = 1.0
-    initializer = Lump(dim=dim, center=orig, velocity=vel)
-    # initializer = Uniform(dim=dim, velocity=vel)
+        vis_timer = IntervalTimer("t_vis", "Time spent visualizing")
+        logmgr.add_quantity(vis_timer)
 
-    from mirgecom.gas_model import GasModel, make_fluid_state
-    gas_model = GasModel(eos=eos)
+    mu = 1.0
+    kappa = 1.0
 
-    def boundary_solution(dcoll, dd_bdry, gas_model, state_minus, **kwargs):
+    top_boundary_temperature = 400
+    bottom_boundary_temperature = 300
+
+    def tramp_2d(x_vec, eos, cv=None, **kwargs):
+        y = x_vec[1]
+        ones = 0*y + 1.0
+        l_y = top_boundary_location - bottom_boundary_location
+        p0 = eos.gas_const() * bottom_boundary_temperature
+        delta_temp = top_boundary_temperature - bottom_boundary_temperature
+        dtdy = delta_temp / l_y
+        temperature_y = bottom_boundary_temperature + dtdy*y
+        mass = p0 / (eos.gas_const() * temperature_y)
+        e0 = p0 / (eos.gamma() - 1.0)
+        velocity = 0 * x_vec
+        energy = e0 * ones
+        momentum = mass * velocity
+        return make_conserved(2, mass=mass, energy=energy, momentum=momentum)
+
+    initializer = tramp_2d
+    gas_model = GasModel(eos=IdealSingleGas(),
+                         transport=SimpleTransport(viscosity=mu,
+                                                   thermal_conductivity=kappa))
+
+    exact = initializer(x_vec=nodes, eos=gas_model.eos)
+
+    def _boundary_state(dcoll, dd_bdry, gas_model, state_minus, **kwargs):
         actx = state_minus.array_context
         bnd_discr = dcoll.discr_from_dd(dd_bdry)
         nodes = actx.thaw(bnd_discr.nodes())
@@ -182,8 +209,14 @@ def main(actx_class, use_esdg=False,
                                             **kwargs), gas_model)
 
     boundaries = {
-        BTAG_ALL: PrescribedFluidBoundary(boundary_state_func=boundary_solution)
-    }
+        BoundaryDomainTag("-1"): PrescribedFluidBoundary(
+            boundary_state_func=_boundary_state),
+        BoundaryDomainTag("+1"): PrescribedFluidBoundary(
+            boundary_state_func=_boundary_state),
+        BoundaryDomainTag("-2"): IsothermalWallBoundary(
+            wall_temperature=bottom_boundary_temperature),
+        BoundaryDomainTag("+2"): IsothermalWallBoundary(
+            wall_temperature=top_boundary_temperature)}
 
     if rst_filename:
         current_t = restart_data["t"]
@@ -194,44 +227,64 @@ def main(actx_class, use_esdg=False,
             logmgr_set_time(logmgr, current_step, current_t)
     else:
         # Set the current state from time 0
-        current_cv = initializer(nodes)
-    current_state = make_fluid_state(current_cv, gas_model)
+        current_cv = exact
+    current_state = make_fluid_state(cv=current_cv, gas_model=gas_model)
 
-    visualizer = make_visualizer(dcoll)
-    initname = initializer.__class__.__name__
-    eosname = eos.__class__.__name__
+    vis_timer = None
+
+    visualizer = make_visualizer(dcoll, order)
+
+    eosname = gas_model.eos.__class__.__name__
     init_message = make_init_message(dim=dim, order=order,
                                      nelements=local_nelements,
                                      global_nelements=global_nelements,
                                      dt=current_dt, t_final=t_final, nstatus=nstatus,
                                      nviz=nviz, cfl=current_cfl,
-                                     constant_cfl=constant_cfl, initname=initname,
+                                     constant_cfl=constant_cfl, initname=casename,
                                      eosname=eosname, casename=casename)
     if rank == 0:
         logger.info(init_message)
 
-    def my_write_viz(step, t, state, dv=None, exact=None, resid=None):
-        if dv is None:
-            dv = eos.dependent_vars(state)
-        if exact is None:
-            exact = initializer(x_vec=nodes, eos=eos, time=t)
-        if resid is None:
-            resid = state - exact
+    def my_write_status(step, t, dt, state, component_errors):
+        from grudge.op import nodal_min, nodal_max
+        dv = state.dv
+        p_min = actx.to_numpy(nodal_min(dcoll, "vol", dv.pressure))
+        p_max = actx.to_numpy(nodal_max(dcoll, "vol", dv.pressure))
+        t_min = actx.to_numpy(nodal_min(dcoll, "vol", dv.temperature))
+        t_max = actx.to_numpy(nodal_max(dcoll, "vol", dv.temperature))
+        if constant_cfl:
+            cfl = current_cfl
+        else:
+            from mirgecom.viscous import get_viscous_cfl
+            cfl = actx.to_numpy(nodal_max(dcoll, "vol",
+                                          get_viscous_cfl(dcoll, dt=current_dt,
+                                                          state=state)))
+
+        if rank == 0:
+            logger.info(f"Step: {step}, T: {t}, DT: {dt}, CFL: {cfl}\n"
+                        f"----- Pressure({p_min}, {p_max})\n"
+                        f"----- Temperature({t_min}, {t_max})\n"
+                        "----- errors="
+                        + ", ".join("%.3g" % en for en in component_errors))
+
+    def my_write_viz(step, t, state, dv):
+        resid = state - exact
         viz_fields = [("cv", state),
                       ("dv", dv),
                       ("exact", exact),
-                      ("residual", resid)]
+                      ("resid", resid)]
+
         from mirgecom.simutil import write_visfile
         write_visfile(dcoll, viz_fields, visualizer, vizname=casename,
-                      step=step, t=t, overwrite=True, vis_timer=vis_timer,
+                      step=step, t=t, overwrite=True,
                       comm=comm)
 
-    def my_write_restart(state, step, t):
+    def my_write_restart(step, t, state):
         rst_fname = rst_pattern.format(cname=casename, step=step, rank=rank)
         if rst_fname != rst_filename:
             rst_data = {
                 "local_mesh": local_mesh,
-                "state": state,
+                "cv": state,
                 "t": t,
                 "step": step,
                 "order": order,
@@ -241,20 +294,34 @@ def main(actx_class, use_esdg=False,
             from mirgecom.restart import write_restart_file
             write_restart_file(actx, rst_data, rst_fname, comm)
 
-    def my_health_check(dv, state, exact):
+    def my_health_check(state, dv, component_errors):
         health_error = False
         from mirgecom.simutil import check_naninf_local, check_range_local
         if check_naninf_local(dcoll, "vol", dv.pressure):
             health_error = True
-            logger.info(f"{rank=}: Invalid pressure data found.")
+            logger.info(f"{rank=}: NANs/Infs in pressure data.")
 
-        if check_range_local(dcoll, "vol", dv.pressure, .9999999999, 1.00000001):
+        if global_reduce(check_range_local(dcoll, "vol", dv.pressure, 86129, 86131),
+                         op="lor"):
             health_error = True
-            logger.info(f"{rank=}: Pressure range violation.")
+            from grudge.op import nodal_max, nodal_min
+            p_min = nodal_min(dcoll, "vol", dv.pressure)
+            p_max = nodal_max(dcoll, "vol", dv.pressure)
+            logger.info(f"Pressure range violation ({p_min=}, {p_max=})")
 
-        from mirgecom.simutil import compare_fluid_solutions
-        component_errors = compare_fluid_solutions(dcoll, state, exact)
-        exittol = .09
+        if check_naninf_local(dcoll, "vol", dv.temperature):
+            health_error = True
+            logger.info(f"{rank=}: NANs/INFs in temperature data.")
+
+        if global_reduce(check_range_local(dcoll, "vol", dv.temperature, 299, 401),
+                         op="lor"):
+            health_error = True
+            from grudge.op import nodal_max, nodal_min
+            t_min = actx.to_numpy(nodal_min(dcoll, "vol", dv.temperature))
+            t_max = actx.to_numpy(nodal_max(dcoll, "vol", dv.temperature))
+            logger.info(f"Temperature range violation ({t_min=}, {t_max=})")
+
+        exittol = .1
         if max(component_errors) > exittol:
             health_error = True
             if rank == 0:
@@ -265,8 +332,9 @@ def main(actx_class, use_esdg=False,
     def my_pre_step(step, t, dt, state):
         fluid_state = make_fluid_state(state, gas_model)
         dv = fluid_state.dv
+
         try:
-            exact = None
+            component_errors = None
 
             if logmgr:
                 logmgr.tick_before()
@@ -278,44 +346,38 @@ def main(actx_class, use_esdg=False,
             do_status = check_step(step=step, interval=nstatus)
 
             if do_health:
-                exact = initializer(x_vec=nodes, eos=eos, time=t)
+                from mirgecom.simutil import compare_fluid_solutions
+                component_errors = compare_fluid_solutions(dcoll, state, exact)
                 health_errors = global_reduce(
-                    my_health_check(dv=dv, state=state, exact=exact), op="lor")
+                    my_health_check(state, dv, component_errors), op="lor")
                 if health_errors:
                     if rank == 0:
                         logger.info("Fluid solution failed health check.")
-                    raise MyRuntimeError("Failed solution health check.")
+                    raise MyRuntimeError("Failed simulation health check.")
 
             if do_restart:
                 my_write_restart(step=step, t=t, state=state)
 
             if do_viz:
-                if exact is None:
-                    exact = initializer(x_vec=nodes, eos=eos, time=t)
-                resid = state - exact
-                my_write_viz(step=step, t=t, dv=dv, state=state, exact=exact,
-                             resid=resid)
+                my_write_viz(step=step, t=t, state=state, dv=dv)
+
+            dt = get_sim_timestep(dcoll, fluid_state, t, dt, current_cfl,
+                                  t_final, constant_cfl)
 
             if do_status:
-                if exact is None:
-                    exact = initializer(x_vec=nodes, eos=eos, time=t)
-                from mirgecom.simutil import compare_fluid_solutions
-                component_errors = compare_fluid_solutions(dcoll, state, exact)
-                status_msg = (
-                    "------- errors="
-                    + ", ".join("%.3g" % en for en in component_errors))
-                if rank == 0:
-                    logger.info(status_msg)
+                if component_errors is None:
+                    from mirgecom.simutil import compare_fluid_solutions
+                    component_errors = compare_fluid_solutions(dcoll, state, exact)
+                my_write_status(step=step, t=t, dt=dt, state=fluid_state,
+                                component_errors=component_errors)
 
         except MyRuntimeError:
             if rank == 0:
                 logger.info("Errors detected; attempting graceful exit.")
-            my_write_viz(step=step, t=t, state=state)
+            my_write_viz(step=step, t=t, state=state, dv=dv)
             my_write_restart(step=step, t=t, state=state)
             raise
 
-        dt = get_sim_timestep(dcoll, fluid_state, t, dt, current_cfl, t_final,
-                              constant_cfl)
         return state, dt
 
     def my_post_step(step, t, dt, state):
@@ -323,15 +385,15 @@ def main(actx_class, use_esdg=False,
         # imo this is a design/scope flaw
         if logmgr:
             set_dt(logmgr, dt)
-            set_sim_state(logmgr, dim, state, eos)
+            set_sim_state(logmgr, dim, state, gas_model.eos)
             logmgr.tick_after()
         return state, dt
 
     def my_rhs(t, state):
         fluid_state = make_fluid_state(state, gas_model)
-        return euler_operator(dcoll, state=fluid_state, time=t,
-                              boundaries=boundaries, gas_model=gas_model,
-                              quadrature_tag=quadrature_tag, use_esdg=use_esdg)
+        return ns_operator(dcoll, boundaries=boundaries, state=fluid_state,
+                           time=t, gas_model=gas_model, use_esdg=use_esdg,
+                           quadrature_tag=quadrature_tag)
 
     current_dt = get_sim_timestep(dcoll, current_state, current_t, current_dt,
                                   current_cfl, t_final, constant_cfl)
@@ -341,18 +403,21 @@ def main(actx_class, use_esdg=False,
                       pre_step_callback=my_pre_step,
                       post_step_callback=my_post_step, dt=current_dt,
                       state=current_state.cv, t=current_t, t_final=t_final)
+    current_state = make_fluid_state(current_cv, gas_model)
 
     # Dump the final data
     if rank == 0:
         logger.info("Checkpointing final state ...")
-
-    current_state = make_fluid_state(current_cv, gas_model)
     final_dv = current_state.dv
-    final_exact = initializer(x_vec=nodes, eos=eos, time=current_t)
-    final_resid = current_state.cv - final_exact
-    my_write_viz(step=current_step, t=current_t, state=current_state.cv, dv=final_dv,
-                 exact=final_exact, resid=final_resid)
-    my_write_restart(step=current_step, t=current_t, state=current_state.cv)
+    final_dt = get_sim_timestep(dcoll, current_state, current_t, current_dt,
+                                current_cfl, t_final, constant_cfl)
+    from mirgecom.simutil import compare_fluid_solutions
+    component_errors = compare_fluid_solutions(dcoll, current_cv, exact)
+
+    my_write_viz(step=current_step, t=current_t, state=current_cv, dv=final_dv)
+    my_write_restart(step=current_step, t=current_t, state=current_cv)
+    my_write_status(step=current_step, t=current_t, dt=final_dt,
+                    state=current_state, component_errors=component_errors)
 
     if logmgr:
         logmgr.close()
@@ -360,25 +425,25 @@ def main(actx_class, use_esdg=False,
         print(actx.tabulate_profiling_data())
 
     finish_tol = 1e-16
-    time_err = current_t - t_final
-    if np.abs(time_err) > finish_tol:
-        raise ValueError(f"Simulation did not finish at expected time {time_err=}.")
+    assert np.abs(current_t - t_final) < finish_tol
 
 
 if __name__ == "__main__":
     import argparse
-    casename = "mass-lump"
+    casename = "hotplate"
     parser = argparse.ArgumentParser(description=f"MIRGE-Com Example: {casename}")
     parser.add_argument("--lazy", action="store_true",
         help="switch to a lazy computation mode")
     parser.add_argument("--profiling", action="store_true",
         help="turn on detailed performance profiling")
+    parser.add_argument("--log", action="store_true", default=True,
+        help="turn on logging")
+    parser.add_argument("--overintegration", action="store_true",
+        help="use overintegration in the RHS computations")
+    parser.add_argument("--esdg", action="store_true",
+        help="use flux-differencing/entropy stable DG for inviscid computations.")
     parser.add_argument("--leap", action="store_true",
         help="use leap timestepper")
-    parser.add_argument("--esdg", action="store_true",
-        help="use entropy-stable Euler operator")
-    parser.add_argument("--overintegration", action="store_true",
-        help="use over-integration")
     parser.add_argument("--numpy", action="store_true",
         help="use numpy-based eager actx.")
     parser.add_argument("--restart_file", help="root name of restart file")
@@ -405,7 +470,7 @@ if __name__ == "__main__":
         rst_filename = args.restart_file
 
     main(actx_class, use_leap=args.leap,
-         casename=casename, rst_filename=rst_filename,
-         use_esdg=args.esdg, use_overintegration=args.esdg or args.overintegration)
+         casename=casename, rst_filename=rst_filename, use_esdg=args.esdg,
+         use_overintegration=args.overintegration or args.esdg)
 
 # vim: foldmethod=marker
