--- examples/lump-mpi-lazy.py
@@ -1,1 +1,0 @@
-lump-mpi.py
