+++ examples/poiseuille-local_dt.py
@@ -1,4 +1,4 @@
-"""Demonstrate a fluid between two hot plates in 2d."""
+"""Demonstrate a planar Poiseuille flow example."""
 
 __copyright__ = """
 Copyright (C) 2020 University of Illinois Board of Trustees
@@ -25,20 +25,20 @@ THE SOFTWARE.
 """
 import logging
 import numpy as np
+from pytools.obj_array import make_obj_array
 from functools import partial
 
+from arraycontext import thaw
 from meshmode.mesh import BTAG_ALL, BTAG_NONE  # noqa
 
+from grudge.eager import EagerDGDiscretization
 from grudge.shortcuts import make_visualizer
 from grudge.dof_desc import BoundaryDomainTag
 
-from mirgecom.discretization import create_discretization_collection
 from mirgecom.fluid import make_conserved
 from mirgecom.navierstokes import ns_operator
-from mirgecom.simutil import (
-    get_sim_timestep,
-    get_box_mesh
-)
+from mirgecom.simutil import get_sim_timestep
+from mirgecom.utils import force_evaluation
 
 from mirgecom.io import make_init_message
 from mirgecom.mpi import mpi_entry_point
@@ -46,7 +46,7 @@ from mirgecom.integrators import rk4_step
 from mirgecom.steppers import advance_state
 from mirgecom.boundary import (
     PrescribedFluidBoundary,
-    IsothermalWallBoundary
+    AdiabaticNoslipWallBoundary
 )
 from mirgecom.transport import SimpleTransport
 from mirgecom.eos import IdealSingleGas
@@ -56,9 +56,8 @@ from mirgecom.euler import extract_vars_for_logging, units_for_logging
 from mirgecom.logging_quantities import (
     initialize_logmgr,
     logmgr_add_many_discretization_quantities,
-    logmgr_add_cl_device_info,
-    logmgr_add_device_memory_usage,
-    set_sim_state
+    logmgr_add_device_name,
+    logmgr_add_device_memory_usage
 )
 
 
@@ -72,8 +71,8 @@ class MyRuntimeError(RuntimeError):
 
 
 @mpi_entry_point
-def main(actx_class, use_esdg=False, use_leap=False, casename=None,
-         use_overintegration=False, rst_filename=None):
+def main(actx_class, use_esdg=False, use_overintegration=False,
+         use_leap=False, casename=None, rst_filename=None):
     """Drive the example."""
     if casename is None:
         casename = "mirgecom"
@@ -96,16 +95,17 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
 
     # timestepping control
     timestepper = rk4_step
-    t_final = 2e-7
-    current_cfl = .1
-    current_dt = 1e-8
+    t_final = 1e-7
+    current_cfl = 0.05
+    current_dt = 1e-10
     current_t = 0
-    constant_cfl = False
+    constant_cfl = True
+    local_dt = True  # XXX declaration necessary for local time stepping.
     current_step = 0
 
     # some i/o frequencies
     nstatus = 1
-    nviz = 1
+    nviz = 100
     nrestart = 100
     nhealth = 1
 
@@ -115,8 +115,8 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
         raise ValueError("This example must be run with dim = 2.")
     left_boundary_location = 0
     right_boundary_location = 0.1
-    bottom_boundary_location = 0
-    top_boundary_location = .02
+    ybottom = 0.
+    ytop = .02
     rst_path = "restart_data/"
     rst_pattern = (
         rst_path + "{cname}-{step:04d}-{rank:04d}.pkl"
@@ -131,30 +131,52 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
         global_nelements = restart_data["global_nelements"]
         assert restart_data["nparts"] == nparts
     else:  # generate the grid from scratch
-        n_refine = 2
-        nels_x = 5 * n_refine
-        nels_y = 3 * n_refine
-        nels_axis = (nels_x, nels_y)
-        box_ll = (left_boundary_location, bottom_boundary_location)
-        box_ur = (right_boundary_location, top_boundary_location)
-        generate_mesh = partial(get_box_mesh, 2, a=box_ll, b=box_ur, n=nels_axis)
+        n_refine = 5
+        npts_x = 10 * n_refine
+        npts_y = 6 * n_refine
+
+        # create a stretched grid to force different grid cell size
+        xx = np.linspace(left_boundary_location, right_boundary_location, npts_x + 1)
+        yy = np.sqrt(np.linspace(0.0, 1.0, npts_y + 1))*(ytop - ybottom)
+        coords = (xx, yy)
+
+        from meshmode.mesh.generation import generate_box_mesh
+        generate_mesh = partial(generate_box_mesh,
+                                axis_coords=coords,
+                                boundary_tag_to_face={
+                                    "-1": ["-x"],
+                                    "+1": ["+x"],
+                                    "-2": ["-y"],
+                                    "+2": ["+y"]})
+
         from mirgecom.simutil import generate_and_distribute_mesh
-        local_mesh, global_nelements = generate_and_distribute_mesh(comm,
-                                                                    generate_mesh)
+        local_mesh, global_nelements = (
+            generate_and_distribute_mesh(comm, generate_mesh))
         local_nelements = local_mesh.nelements
 
-    order = 1
-    dcoll = create_discretization_collection(actx, local_mesh, order=order)
-    nodes = actx.thaw(dcoll.nodes())
+    from grudge.dof_desc import DISCR_TAG_BASE, DISCR_TAG_QUAD
+    from meshmode.discretization.poly_element import \
+        default_simplex_group_factory, QuadratureSimplexGroupFactory
 
-    from grudge.dof_desc import DISCR_TAG_QUAD
+    order = 2
+    dcoll = EagerDGDiscretization(
+        actx, local_mesh,
+        discr_tag_to_group_factory={
+            DISCR_TAG_BASE: default_simplex_group_factory(
+                base_dim=local_mesh.dim, order=order),
+            DISCR_TAG_QUAD: QuadratureSimplexGroupFactory(2*order + 1)
+        },
+        mpi_communicator=comm
+    )
+    nodes = thaw(dcoll.nodes(), actx)
+
     if use_overintegration:
         quadrature_tag = DISCR_TAG_QUAD
     else:
         quadrature_tag = None
 
     if logmgr:
-        logmgr_add_cl_device_info(logmgr, queue)
+        logmgr_add_device_name(logmgr, queue)
         logmgr_add_device_memory_usage(logmgr, queue)
         logmgr_add_many_discretization_quantities(logmgr, dcoll, dim,
                              extract_vars_for_logging, units_for_logging)
@@ -162,10 +184,6 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
         logmgr.add_watches([
             ("step.max", "step = {value}, "),
             ("t_sim.max", "sim time: {value:1.6e} s\n"),
-            ("min_pressure", "------- P (min, max) (Pa) = ({value:1.9e}, "),
-            ("max_pressure",    "{value:1.9e})\n"),
-            ("min_temperature", "------- T (min, max) (K) = ({value:1.9e}, "),
-            ("max_temperature",    "{value:1.9e})\n"),
             ("t_step.max", "------- step walltime: {value:6g} s, "),
             ("t_log.max", "log walltime: {value:6g} s")
         ])
@@ -173,53 +191,63 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
         vis_timer = IntervalTimer("t_vis", "Time spent visualizing")
         logmgr.add_quantity(vis_timer)
 
+    base_pressure = 100000.0
+    pressure_ratio = 1.001
     mu = 1.0
-    kappa = 1.0
 
-    top_boundary_temperature = 400
-    bottom_boundary_temperature = 300
-
-    def tramp_2d(x_vec, eos, cv=None, **kwargs):
+    def poiseuille_2d(x_vec, eos, cv=None, **kwargs):
         y = x_vec[1]
-        ones = 0*y + 1.0
-        l_y = top_boundary_location - bottom_boundary_location
-        p0 = eos.gas_const() * bottom_boundary_temperature
-        delta_temp = top_boundary_temperature - bottom_boundary_temperature
-        dtdy = delta_temp / l_y
-        temperature_y = bottom_boundary_temperature + dtdy*y
-        mass = p0 / (eos.gas_const() * temperature_y)
-        e0 = p0 / (eos.gamma() - 1.0)
-        velocity = 0 * x_vec
-        energy = e0 * ones
-        momentum = mass * velocity
-        return make_conserved(2, mass=mass, energy=energy, momentum=momentum)
+        x = x_vec[0]
+        x0 = left_boundary_location
+        xmax = right_boundary_location
+        xlen = xmax - x0
+        p_low = base_pressure
+        p_hi = pressure_ratio*base_pressure
+        dp = p_hi - p_low
+        dpdx = dp/xlen
+        h = ytop - ybottom
+        u_x = dpdx*y*(h - y)/(2*mu)
+        p_x = p_hi - dpdx*x
+        rho = 1.0
+        mass = 0*x + rho
+        u_y = 0*x
+        velocity = make_obj_array([u_x, u_y])
+        ke = .5*np.dot(velocity, velocity)*mass
+        gamma = eos.gamma()
+        if cv is not None:
+            mass = cv.mass
+            vel = cv.velocity
+            ke = .5*np.dot(vel, vel)*mass
 
-    initializer = tramp_2d
+        rho_e = p_x/(gamma-1) + ke
+        return make_conserved(2, mass=mass, energy=rho_e,
+                              momentum=mass*velocity)
+
+    initializer = poiseuille_2d
     gas_model = GasModel(eos=IdealSingleGas(),
-                         transport=SimpleTransport(viscosity=mu,
-                                                   thermal_conductivity=kappa))
-
+                         transport=SimpleTransport(viscosity=mu))
     exact = initializer(x_vec=nodes, eos=gas_model.eos)
 
-    def _boundary_state(dcoll, dd_bdry, gas_model, state_minus, **kwargs):
+    def _boundary_solution(dcoll, dd_bdry, gas_model, state_minus, **kwargs):
         actx = state_minus.array_context
         bnd_discr = dcoll.discr_from_dd(dd_bdry)
-        nodes = actx.thaw(bnd_discr.nodes())
+        nodes = thaw(bnd_discr.nodes(), actx)
         return make_fluid_state(initializer(x_vec=nodes, eos=gas_model.eos,
-                                            **kwargs), gas_model)
+                                            cv=state_minus.cv, **kwargs), gas_model)
 
     boundaries = {
-        BoundaryDomainTag("-1"): PrescribedFluidBoundary(
-            boundary_state_func=_boundary_state),
-        BoundaryDomainTag("+1"): PrescribedFluidBoundary(
-            boundary_state_func=_boundary_state),
-        BoundaryDomainTag("-2"): IsothermalWallBoundary(
-            wall_temperature=bottom_boundary_temperature),
-        BoundaryDomainTag("+2"): IsothermalWallBoundary(
-            wall_temperature=top_boundary_temperature)}
+        BoundaryDomainTag("-1"):
+            PrescribedFluidBoundary(boundary_state_func=_boundary_solution),
+        BoundaryDomainTag("+1"):
+            PrescribedFluidBoundary(boundary_state_func=_boundary_solution),
+        BoundaryDomainTag("-2"): AdiabaticNoslipWallBoundary(),
+        BoundaryDomainTag("+2"): AdiabaticNoslipWallBoundary()}
 
     if rst_filename:
-        current_t = restart_data["t"]
+        if local_dt:
+            current_t = restart_data["step"]
+        else:
+            current_t = restart_data["t"]
         current_step = restart_data["step"]
         current_cv = restart_data["cv"]
         if logmgr:
@@ -228,6 +256,7 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
     else:
         # Set the current state from time 0
         current_cv = exact
+
     current_state = make_fluid_state(cv=current_cv, gas_model=gas_model)
 
     vis_timer = None
@@ -246,23 +275,14 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
         logger.info(init_message)
 
     def my_write_status(step, t, dt, state, component_errors):
-        from grudge.op import nodal_min, nodal_max
         dv = state.dv
+        from grudge.op import nodal_min, nodal_max
         p_min = actx.to_numpy(nodal_min(dcoll, "vol", dv.pressure))
         p_max = actx.to_numpy(nodal_max(dcoll, "vol", dv.pressure))
         t_min = actx.to_numpy(nodal_min(dcoll, "vol", dv.temperature))
         t_max = actx.to_numpy(nodal_max(dcoll, "vol", dv.temperature))
-        if constant_cfl:
-            cfl = current_cfl
-        else:
-            from mirgecom.viscous import get_viscous_cfl
-            cfl = actx.to_numpy(nodal_max(dcoll, "vol",
-                                          get_viscous_cfl(dcoll, dt=current_dt,
-                                                          state=state)))
-
         if rank == 0:
-            logger.info(f"Step: {step}, T: {t}, DT: {dt}, CFL: {cfl}\n"
-                        f"----- Pressure({p_min}, {p_max})\n"
+            logger.info(f"----- Pressure({p_min}, {p_max})\n"
                         f"----- Temperature({t_min}, {t_max})\n"
                         "----- errors="
                         + ", ".join("%.3g" % en for en in component_errors))
@@ -271,13 +291,12 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
         resid = state - exact
         viz_fields = [("cv", state),
                       ("dv", dv),
-                      ("exact", exact),
+                      ("poiseuille", exact),
                       ("resid", resid)]
 
         from mirgecom.simutil import write_visfile
         write_visfile(dcoll, viz_fields, visualizer, vizname=casename,
-                      step=step, t=t, overwrite=True,
-                      comm=comm)
+                      step=step, t=t, overwrite=True)
 
     def my_write_restart(step, t, state):
         rst_fname = rst_pattern.format(cname=casename, step=step, rank=rank)
@@ -301,19 +320,19 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
             health_error = True
             logger.info(f"{rank=}: NANs/Infs in pressure data.")
 
-        if global_reduce(check_range_local(dcoll, "vol", dv.pressure, 86129, 86131),
-                         op="lor"):
+        if global_reduce(check_range_local(dcoll, "vol", dv.pressure, 9.999e4,
+                                           1.00101e5), op="lor"):
             health_error = True
             from grudge.op import nodal_max, nodal_min
-            p_min = nodal_min(dcoll, "vol", dv.pressure)
-            p_max = nodal_max(dcoll, "vol", dv.pressure)
+            p_min = actx.to_numpy(nodal_min(dcoll, "vol", dv.pressure))
+            p_max = actx.to_numpy(nodal_max(dcoll, "vol", dv.pressure))
             logger.info(f"Pressure range violation ({p_min=}, {p_max=})")
 
         if check_naninf_local(dcoll, "vol", dv.temperature):
             health_error = True
             logger.info(f"{rank=}: NANs/INFs in temperature data.")
 
-        if global_reduce(check_range_local(dcoll, "vol", dv.temperature, 299, 401),
+        if global_reduce(check_range_local(dcoll, "vol", dv.temperature, 348, 350),
                          op="lor"):
             health_error = True
             from grudge.op import nodal_max, nodal_min
@@ -330,8 +349,16 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
         return health_error
 
     def my_pre_step(step, t, dt, state):
-        fluid_state = make_fluid_state(state, gas_model)
+        fluid_state = make_fluid_state(cv=state, gas_model=gas_model)
         dv = fluid_state.dv
+
+        if constant_cfl:
+            dt = get_sim_timestep(dcoll, fluid_state, t, dt, current_cfl,
+                                           t_final, constant_cfl, local_dt)
+        if local_dt:
+            t = force_evaluation(actx, t)
+            dt = force_evaluation(actx, get_sim_timestep(dcoll, fluid_state, t, dt,
+                 current_cfl, constant_cfl=constant_cfl, local_dt=local_dt))
 
         try:
             component_errors = None
@@ -361,10 +388,7 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
             if do_viz:
                 my_write_viz(step=step, t=t, state=state, dv=dv)
 
-            dt = get_sim_timestep(dcoll, fluid_state, t, dt, current_cfl,
-                                  t_final, constant_cfl)
-
-            if do_status:
+            if do_status:  # needed because logging fails to make output
                 if component_errors is None:
                     from mirgecom.simutil import compare_fluid_solutions
                     component_errors = compare_fluid_solutions(dcoll, state, exact)
@@ -381,41 +405,49 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
         return state, dt
 
     def my_post_step(step, t, dt, state):
-        # Logmgr needs to know about EOS, dt, dim?
-        # imo this is a design/scope flaw
         if logmgr:
-            set_dt(logmgr, dt)
-            set_sim_state(logmgr, dim, state, gas_model.eos)
+            if local_dt:
+                set_dt(logmgr, 1.0)
+            else:
+                set_dt(logmgr, dt)
             logmgr.tick_after()
         return state, dt
 
     def my_rhs(t, state):
         fluid_state = make_fluid_state(state, gas_model)
-        return ns_operator(dcoll, boundaries=boundaries, state=fluid_state,
-                           time=t, gas_model=gas_model, use_esdg=use_esdg,
+        return ns_operator(dcoll, gas_model=gas_model, boundaries=boundaries,
+                           state=fluid_state, time=t, use_esdg=use_esdg,
                            quadrature_tag=quadrature_tag)
 
     current_dt = get_sim_timestep(dcoll, current_state, current_t, current_dt,
-                                  current_cfl, t_final, constant_cfl)
+                 current_cfl, constant_cfl=constant_cfl, local_dt=local_dt)
+    if local_dt:
+        current_dt = force_evaluation(actx, current_dt)
+
+        current_t = current_t + current_dt*0.0
+        current_t = force_evaluation(actx, current_t)
 
     current_step, current_t, current_cv = \
         advance_state(rhs=my_rhs, timestepper=timestepper,
                       pre_step_callback=my_pre_step,
                       post_step_callback=my_post_step, dt=current_dt,
-                      state=current_state.cv, t=current_t, t_final=t_final)
-    current_state = make_fluid_state(current_cv, gas_model)
+                      state=current_state.cv, t=current_t, t_final=t_final,
+                      max_steps=20, local_dt=local_dt,
+                      istep=current_step)
+
+    current_state = make_fluid_state(cv=current_cv, gas_model=gas_model)
 
     # Dump the final data
     if rank == 0:
         logger.info("Checkpointing final state ...")
     final_dv = current_state.dv
     final_dt = get_sim_timestep(dcoll, current_state, current_t, current_dt,
-                                current_cfl, t_final, constant_cfl)
+                 current_cfl, constant_cfl=constant_cfl, local_dt=local_dt)
     from mirgecom.simutil import compare_fluid_solutions
-    component_errors = compare_fluid_solutions(dcoll, current_cv, exact)
+    component_errors = compare_fluid_solutions(dcoll, current_state.cv, exact)
 
-    my_write_viz(step=current_step, t=current_t, state=current_cv, dv=final_dv)
-    my_write_restart(step=current_step, t=current_t, state=current_cv)
+    my_write_viz(step=current_step, t=current_t, state=current_state.cv, dv=final_dv)
+    my_write_restart(step=current_step, t=current_t, state=current_state)
     my_write_status(step=current_step, t=current_t, dt=final_dt,
                     state=current_state, component_errors=component_errors)
 
@@ -424,26 +456,25 @@ def main(actx_class, use_esdg=False, use_leap=False, casename=None,
     elif use_profiling:
         print(actx.tabulate_profiling_data())
 
-    finish_tol = 1e-16
-    assert np.abs(current_t - t_final) < finish_tol
+    exit()
 
 
 if __name__ == "__main__":
     import argparse
-    casename = "hotplate"
+    casename = "poiseuille"
     parser = argparse.ArgumentParser(description=f"MIRGE-Com Example: {casename}")
+    parser.add_argument("--overintegration", action="store_true",
+        help="use overintegration in the RHS computations")
     parser.add_argument("--lazy", action="store_true",
         help="switch to a lazy computation mode")
     parser.add_argument("--profiling", action="store_true",
         help="turn on detailed performance profiling")
     parser.add_argument("--log", action="store_true", default=True,
         help="turn on logging")
-    parser.add_argument("--overintegration", action="store_true",
-        help="use overintegration in the RHS computations")
-    parser.add_argument("--esdg", action="store_true",
-        help="use flux-differencing/entropy stable DG for inviscid computations.")
     parser.add_argument("--leap", action="store_true",
         help="use leap timestepper")
+    parser.add_argument("--esdg", action="store_true",
+        help="use flux-differencing/entropy stable DG for inviscid computations.")
     parser.add_argument("--numpy", action="store_true",
         help="use numpy-based eager actx.")
     parser.add_argument("--restart_file", help="root name of restart file")
@@ -469,8 +500,8 @@ if __name__ == "__main__":
     if args.restart_file:
         rst_filename = args.restart_file
 
-    main(actx_class, use_leap=args.leap,
-         casename=casename, rst_filename=rst_filename, use_esdg=args.esdg,
-         use_overintegration=args.overintegration or args.esdg)
+    main(actx_class, use_leap=args.leap, use_esdg=args.esdg,
+         use_overintegration=args.overintegration or args.esdg,
+         casename=casename, rst_filename=rst_filename)
 
 # vim: foldmethod=marker
