--- examples/poiseuille-mpi-lazy.py
@@ -1,1 +1,0 @@
-poiseuille-mpi.py
