+++ examples/poiseuille.py
