--- examples/pulse-mpi-lazy.py
@@ -1,1 +1,0 @@
-pulse-mpi.py
