+++ examples/pulse.py
@@ -116,7 +116,7 @@ def main(actx_class, use_esdg=False,
     # some i/o frequencies
     nstatus = 1
     nrestart = 5
-    nviz = 10
+    nviz = 100
     nhealth = 1
 
     dim = 2
