+++ examples/scalar-advdiff.py
