--- examples/scalar-lump-mpi-lazy.py
@@ -1,1 +1,0 @@
-scalar-lump-mpi.py
