--- examples/sod-mpi-lazy.py
@@ -1,1 +1,0 @@
-sod-mpi.py
