+++ examples/thermally-coupled.py
@@ -112,9 +112,9 @@ def main(actx_class, use_esdg=False, use_overintegration=False,
         timestepper = RK4MethodBuilder("state")
     else:
         timestepper = rk4_step
-    t_final = 2e-7
+    t_final = 2e-9
     current_cfl = 1.0
-    current_dt = 1e-8
+    current_dt = 1e-10
     current_t = 0
     constant_cfl = False
 
@@ -124,7 +124,7 @@ def main(actx_class, use_esdg=False, use_overintegration=False,
     # some i/o frequencies
     nstatus = 1
     nrestart = 500
-    nviz = 25
+    nviz = 100
     nhealth = 1
 
     dim = 2
