--- examples/vortex-mpi-lazy.py
@@ -1,1 +1,0 @@
-vortex-mpi.py
