+++ examples/vortex.py
@@ -1,4 +1,4 @@
-"""Demonstrate simple gas mixture with Pyrometheus."""
+"""Demonstrate the isentropic vortex example."""
 
 __copyright__ = """
 Copyright (C) 2020 University of Illinois Board of Trustees
@@ -29,30 +29,26 @@ from functools import partial
 
 from meshmode.mesh import BTAG_ALL, BTAG_NONE  # noqa
 from grudge.shortcuts import make_visualizer
-from grudge.dof_desc import DISCR_TAG_QUAD
 
 from mirgecom.discretization import create_discretization_collection
 from mirgecom.euler import euler_operator
 from mirgecom.simutil import (
     get_sim_timestep,
-    generate_and_distribute_mesh
+    generate_and_distribute_mesh,
+    check_step
 )
 from mirgecom.io import make_init_message
 from mirgecom.mpi import mpi_entry_point
 
 from mirgecom.integrators import rk4_step
 from mirgecom.steppers import advance_state
-from mirgecom.boundary import (
-    PrescribedFluidBoundary,
-    AdiabaticSlipBoundary
-)
-from mirgecom.initializers import MixtureInitializer
-from mirgecom.eos import PyrometheusMixture
-
-import cantera
-
+from mirgecom.boundary import PrescribedFluidBoundary
+from mirgecom.initializers import Vortex2D
+from mirgecom.eos import IdealSingleGas
+from mirgecom.gas_model import GasModel, make_fluid_state
 from logpyle import IntervalTimer, set_dt
 from mirgecom.euler import extract_vars_for_logging, units_for_logging
+
 from mirgecom.logging_quantities import (
     initialize_logmgr,
     logmgr_add_many_discretization_quantities,
@@ -71,17 +67,16 @@ class MyRuntimeError(RuntimeError):
 
 
 @mpi_entry_point
-def main(actx_class, use_esdg=False,
-         use_leap=False, casename=None, rst_filename=None,
-         log_dependent=False, use_overintegration=False):
-    """Drive example."""
+def main(actx_class, use_overintegration=False, use_esdg=False,
+         use_leap=False, casename=None, rst_filename=None):
+    """Drive the example."""
     if casename is None:
         casename = "mirgecom"
 
     from mpi4py import MPI
     comm = MPI.COMM_WORLD
     rank = comm.Get_rank()
-    nparts = comm.Get_size()
+    num_parts = comm.Get_size()
 
     from mirgecom.simutil import global_reduce as _global_reduce
     global_reduce = partial(_global_reduce, comm=comm)
@@ -95,26 +90,28 @@ def main(actx_class, use_esdg=False,
     use_profiling = actx_class_is_profiling(actx_class)
 
     # timestepping control
+    current_step = 0
     if use_leap:
         from leap.rk import RK4MethodBuilder
         timestepper = RK4MethodBuilder("state")
     else:
         timestepper = rk4_step
-
-    t_final = 1e-8
+    t_final = 0.01
     current_cfl = 1.0
-    current_dt = 1e-9
+    current_dt = .001
     current_t = 0
-    current_step = 0
     constant_cfl = False
 
     # some i/o frequencies
+    nrestart = 10
     nstatus = 1
-    nhealth = 1
-    nrestart = 5
-    nviz = 1
+    nviz = 100
+    nhealth = 10
 
     dim = 2
+    if dim != 2:
+        raise ValueError("This example must be run with dim = 2.")
+
     rst_path = "restart_data/"
     rst_pattern = (
         rst_path + "{cname}-{step:04d}-{rank:04d}.pkl"
@@ -126,7 +123,7 @@ def main(actx_class, use_esdg=False,
         local_mesh = restart_data["local_mesh"]
         local_nelements = local_mesh.nelements
         global_nelements = restart_data["global_nelements"]
-        assert restart_data["num_parts"] == nparts
+        assert restart_data["num_parts"] == num_parts
     else:  # generate the grid from scratch
         nel_1d = 16
         box_ll = -5.0
@@ -142,6 +139,7 @@ def main(actx_class, use_esdg=False,
     dcoll = create_discretization_collection(actx, local_mesh, order=order)
     nodes = actx.thaw(dcoll.nodes())
 
+    from grudge.dof_desc import DISCR_TAG_QUAD
     if use_overintegration:
         quadrature_tag = DISCR_TAG_QUAD
     else:
@@ -152,6 +150,8 @@ def main(actx_class, use_esdg=False,
     if logmgr:
         logmgr_add_cl_device_info(logmgr, queue)
         logmgr_add_device_memory_usage(logmgr, queue)
+        logmgr_add_many_discretization_quantities(logmgr, dcoll, dim,
+                             extract_vars_for_logging, units_for_logging)
 
         vis_timer = IntervalTimer("t_vis", "Time spent visualizing")
         logmgr.add_quantity(vis_timer)
@@ -159,82 +159,54 @@ def main(actx_class, use_esdg=False,
         logmgr.add_watches([
             ("step.max", "step = {value}, "),
             ("t_sim.max", "sim time: {value:1.6e} s\n"),
+            ("min_pressure", "------- P (min, max) (Pa) = ({value:1.9e}, "),
+            ("max_pressure",    "{value:1.9e})\n"),
             ("t_step.max", "------- step walltime: {value:6g} s, "),
             ("t_log.max", "log walltime: {value:6g} s")
         ])
 
-        if log_dependent:
-            logmgr_add_many_discretization_quantities(logmgr, dcoll, dim,
-                                                      extract_vars_for_logging,
-                                                      units_for_logging)
-            logmgr.add_watches([
-                ("min_pressure", "\n------- P (min, max) (Pa) = ({value:1.9e}, "),
-                ("max_pressure",    "{value:1.9e})\n"),
-                ("min_temperature", "------- T (min, max) (K)  = ({value:7g}, "),
-                ("max_temperature",    "{value:7g})\n")])
+        try:
+            logmgr.add_watches(["memory_usage_python.max", "memory_usage_gpu.max"])
+        except KeyError:
+            pass
 
-    # Pyrometheus initialization
-    from mirgecom.mechanisms import get_mechanism_input
-    mech_input = get_mechanism_input("uiuc_7sp")
-    sol = cantera.Solution(name="gas", yaml=mech_input)
-    from mirgecom.thermochemistry import get_pyrometheus_wrapper_class_from_cantera
-    pyrometheus_mechanism = \
-        get_pyrometheus_wrapper_class_from_cantera(sol)(actx.np)
+        if use_profiling:
+            logmgr.add_watches(["multiply_time.max"])
 
-    nspecies = pyrometheus_mechanism.num_species
-    eos = PyrometheusMixture(pyrometheus_mechanism, temperature_guess=300)
-    from mirgecom.gas_model import GasModel, make_fluid_state
+    # soln setup and init
+    eos = IdealSingleGas()
+    vel = np.zeros(shape=(dim,))
+    orig = np.zeros(shape=(dim,))
+    vel[:dim] = 1.0
+    initializer = Vortex2D(center=orig, velocity=vel)
     gas_model = GasModel(eos=eos)
-    from pytools.obj_array import make_obj_array
-
-    y0s = np.zeros(shape=(nspecies,))
-    for i in range(nspecies-1):
-        y0s[i] = 1.0 / (10.0 ** (i + 1))
-    spec_sum = sum([y0s[i] for i in range(nspecies-1)])
-    y0s[nspecies-1] = 1.0 - spec_sum
-
-    # Mixture defaults to STP (p, T) = (1atm, 300K)
-    velocity = np.zeros(shape=(dim,)) + 1.0
-    initializer = MixtureInitializer(dim=dim, nspecies=nspecies,
-                                     massfractions=y0s, velocity=velocity)
 
     def boundary_solution(dcoll, dd_bdry, gas_model, state_minus, **kwargs):
         actx = state_minus.array_context
         bnd_discr = dcoll.discr_from_dd(dd_bdry)
         nodes = actx.thaw(bnd_discr.nodes())
         return make_fluid_state(initializer(x_vec=nodes, eos=gas_model.eos,
-                                            **kwargs), gas_model,
-                                temperature_seed=state_minus.temperature)
+                                            **kwargs), gas_model)
 
-    if False:
-        my_boundary = AdiabaticSlipBoundary()
-        boundaries = {BTAG_ALL: my_boundary}
-    else:
-        boundaries = {
-            BTAG_ALL: PrescribedFluidBoundary(boundary_state_func=boundary_solution)
-        }
+    boundaries = {
+        BTAG_ALL: PrescribedFluidBoundary(boundary_state_func=boundary_solution)
+    }
 
     if rst_filename:
         current_t = restart_data["t"]
         current_step = restart_data["step"]
         current_cv = restart_data["cv"]
-        tseed = restart_data["temperature_seed"]
         if logmgr:
             from mirgecom.logging_quantities import logmgr_set_time
             logmgr_set_time(logmgr, current_step, current_t)
     else:
         # Set the current state from time 0
-        current_cv = initializer(x_vec=nodes, eos=eos)
-        tseed = 300.0
+        current_cv = initializer(nodes)
 
-    def get_fluid_state(cv, tseed):
-        return make_fluid_state(cv=cv, gas_model=gas_model,
-                                temperature_seed=tseed)
-
-    construct_fluid_state = actx.compile(get_fluid_state)
-    current_state = construct_fluid_state(current_cv, tseed)
+    current_state = make_fluid_state(current_cv, gas_model)
 
     visualizer = make_visualizer(dcoll)
+
     initname = initializer.__class__.__name__
     eosname = eos.__class__.__name__
     init_message = make_init_message(dim=dim, order=order,
@@ -247,67 +219,61 @@ def main(actx_class, use_esdg=False,
     if rank == 0:
         logger.info(init_message)
 
-    def my_write_status(component_errors, dv=None):
-        status_msg = (
-            "------- errors="
-            + ", ".join("%.3g" % en for en in component_errors))
-        if ((dv is not None) and (not log_dependent)):
-            temp = dv.temperature
-            press = dv.pressure
-
-            from grudge.op import nodal_min_loc, nodal_max_loc
-            tmin = global_reduce(actx.to_numpy(nodal_min_loc(dcoll, "vol", temp)),
-                                 op="min")
-            tmax = global_reduce(actx.to_numpy(nodal_max_loc(dcoll, "vol", temp)),
-                                 op="max")
-            pmin = global_reduce(actx.to_numpy(nodal_min_loc(dcoll, "vol", press)),
-                                 op="min")
-            pmax = global_reduce(actx.to_numpy(nodal_max_loc(dcoll, "vol", press)),
-                                 op="max")
-            dv_status_msg = f"\nP({pmin}, {pmax}), T({tmin}, {tmax})"
-            status_msg = status_msg + dv_status_msg
-
+    def my_write_status(state, component_errors, cfl=None):
+        if cfl is None:
+            if constant_cfl:
+                cfl = current_cfl
+            else:
+                from grudge.op import nodal_max
+                from mirgecom.inviscid import get_inviscid_cfl
+                cfl = actx.to_numpy(
+                    nodal_max(
+                        dcoll, "vol",
+                        get_inviscid_cfl(dcoll, state, current_dt)))[()]
         if rank == 0:
-            logger.info(status_msg)
-        if rank == 0:
-            logger.info(status_msg)
+            logger.info(
+                f"------ {cfl=}\n"
+                "------- errors="
+                + ", ".join("%.3g" % en for en in component_errors))
 
-    def my_write_viz(step, t, state, dv, exact=None, resid=None):
+    def my_write_viz(step, t, state, dv=None, exact=None, resid=None):
         if exact is None:
             exact = initializer(x_vec=nodes, eos=eos, time=t)
         if resid is None:
             resid = state - exact
-        viz_fields = [("cv", state), ("dv", dv)]
+        viz_fields = [("cv", state),
+                      ("dv", dv),
+                      ("exact", exact),
+                      ("residual", resid)]
         from mirgecom.simutil import write_visfile
         write_visfile(dcoll, viz_fields, visualizer, vizname=casename,
                       step=step, t=t, overwrite=True, vis_timer=vis_timer,
                       comm=comm)
 
-    def my_write_restart(step, t, state, tseed):
+    def my_write_restart(step, t, state):
         rst_fname = rst_pattern.format(cname=casename, step=step, rank=rank)
         if rst_fname != rst_filename:
             rst_data = {
                 "local_mesh": local_mesh,
                 "cv": state,
-                "temperature_seed": tseed,
                 "t": t,
                 "step": step,
                 "order": order,
                 "global_nelements": global_nelements,
-                "num_parts": nparts
+                "num_parts": num_parts
             }
             from mirgecom.restart import write_restart_file
             write_restart_file(actx, rst_data, rst_fname, comm)
 
-    def my_health_check(dv, component_errors):
+    def my_health_check(pressure, component_errors):
         health_error = False
         from mirgecom.simutil import check_naninf_local, check_range_local
-        if check_naninf_local(dcoll, "vol", dv.pressure) \
-           or check_range_local(dcoll, "vol", dv.pressure, 1e5, 1.1e5):
+        if check_naninf_local(dcoll, "vol", pressure) \
+           or check_range_local(dcoll, "vol", pressure, .2, 1.02):
             health_error = True
             logger.info(f"{rank=}: Invalid pressure data found.")
 
-        exittol = .09
+        exittol = .1
         if max(component_errors) > exittol:
             health_error = True
             if rank == 0:
@@ -316,8 +282,8 @@ def main(actx_class, use_esdg=False,
         return health_error
 
     def my_pre_step(step, t, dt, state):
-        cv, tseed = state
-        fluid_state = construct_fluid_state(cv, tseed)
+        fluid_state = make_fluid_state(state, gas_model)
+        cv = fluid_state.cv
         dv = fluid_state.dv
 
         try:
@@ -327,7 +293,6 @@ def main(actx_class, use_esdg=False,
             if logmgr:
                 logmgr.tick_before()
 
-            from mirgecom.simutil import check_step
             do_viz = check_step(step=step, interval=nviz)
             do_restart = check_step(step=step, interval=nrestart)
             do_health = check_step(step=step, interval=nhealth)
@@ -338,14 +303,22 @@ def main(actx_class, use_esdg=False,
                 from mirgecom.simutil import compare_fluid_solutions
                 component_errors = compare_fluid_solutions(dcoll, cv, exact)
                 health_errors = global_reduce(
-                    my_health_check(dv, component_errors), op="lor")
+                    my_health_check(dv.pressure, component_errors), op="lor")
                 if health_errors:
                     if rank == 0:
                         logger.info("Fluid solution failed health check.")
                     raise MyRuntimeError("Failed simulation health check.")
 
             if do_restart:
-                my_write_restart(step=step, t=t, state=cv, tseed=tseed)
+                my_write_restart(step=step, t=t, state=cv)
+
+            if do_status:
+                if component_errors is None:
+                    if exact is None:
+                        exact = initializer(x_vec=nodes, eos=eos, time=t)
+                    from mirgecom.simutil import compare_fluid_solutions
+                    component_errors = compare_fluid_solutions(dcoll, cv, exact)
+                my_write_status(fluid_state, component_errors)
 
             if do_viz:
                 if exact is None:
@@ -354,19 +327,11 @@ def main(actx_class, use_esdg=False,
                 my_write_viz(step=step, t=t, state=cv, dv=dv, exact=exact,
                              resid=resid)
 
-            if do_status:
-                if component_errors is None:
-                    if exact is None:
-                        exact = initializer(x_vec=nodes, eos=eos, time=t)
-                    from mirgecom.simutil import compare_fluid_solutions
-                    component_errors = compare_fluid_solutions(dcoll, cv, exact)
-                my_write_status(component_errors, dv=dv)
-
         except MyRuntimeError:
             if rank == 0:
                 logger.info("Errors detected; attempting graceful exit.")
-            my_write_viz(step=step, t=t, state=cv, dv=dv)
-            my_write_restart(step=step, t=t, state=cv, tseed=tseed)
+            my_write_viz(step=step, t=t, state=cv)
+            my_write_restart(step=step, t=t, state=cv)
             raise
 
         dt = get_sim_timestep(dcoll, fluid_state, t, dt, current_cfl, t_final,
@@ -374,50 +339,40 @@ def main(actx_class, use_esdg=False,
         return state, dt
 
     def my_post_step(step, t, dt, state):
-        cv, tseed = state
-        fluid_state = construct_fluid_state(cv, tseed)
-        tseed = fluid_state.temperature
         # Logmgr needs to know about EOS, dt, dim?
         # imo this is a design/scope flaw
         if logmgr:
             set_dt(logmgr, dt)
-            set_sim_state(logmgr, dim, cv, eos)
+            set_sim_state(logmgr, dim, state, eos)
             logmgr.tick_after()
-        return make_obj_array([fluid_state.cv, tseed]), dt
+        return state, dt
 
     def my_rhs(t, state):
-        cv, tseed = state
-        fluid_state = make_fluid_state(cv, gas_model, temperature_seed=tseed)
-        return make_obj_array(
-            [euler_operator(dcoll, state=fluid_state, time=t,
-                            boundaries=boundaries, gas_model=gas_model,
-                            quadrature_tag=quadrature_tag, use_esdg=use_esdg),
-             0*tseed])
+        fluid_state = make_fluid_state(state, gas_model)
+        return euler_operator(dcoll, state=fluid_state, time=t,
+                              boundaries=boundaries, gas_model=gas_model,
+                              quadrature_tag=quadrature_tag, use_esdg=use_esdg)
 
     current_dt = get_sim_timestep(dcoll, current_state, current_t, current_dt,
                                   current_cfl, t_final, constant_cfl)
 
-    current_step, current_t, advanced_state = \
+    current_step, current_t, current_cv = \
         advance_state(rhs=my_rhs, timestepper=timestepper,
                       pre_step_callback=my_pre_step,
                       post_step_callback=my_post_step, dt=current_dt,
-                      state=make_obj_array([current_state.cv,
-                                            current_state.temperature]),
-                      t=current_t, t_final=t_final)
+                      state=current_state.cv, t=current_t, t_final=t_final)
 
     # Dump the final data
     if rank == 0:
         logger.info("Checkpointing final state ...")
 
-    current_cv, tseed = advanced_state
-    current_state = make_fluid_state(current_cv, gas_model, temperature_seed=tseed)
+    current_state = make_fluid_state(current_cv, gas_model)
     final_dv = current_state.dv
     final_exact = initializer(x_vec=nodes, eos=eos, time=current_t)
     final_resid = current_state.cv - final_exact
-    my_write_viz(step=current_step, t=current_t, state=current_cv, dv=final_dv,
+    my_write_viz(step=current_step, t=current_t, state=current_state.cv, dv=final_dv,
                  exact=final_exact, resid=final_resid)
-    my_write_restart(step=current_step, t=current_t, state=current_state.cv,
-                     tseed=tseed)
+    my_write_restart(step=current_step, t=current_t, state=current_state.cv)
 
     if logmgr:
         logmgr.close()
@@ -430,27 +385,25 @@ def main(actx_class, use_esdg=False,
 
 if __name__ == "__main__":
     import argparse
-    casename = "uiuc-mixture"
+    casename = "vortex"
     parser = argparse.ArgumentParser(description=f"MIRGE-Com Example: {casename}")
     parser.add_argument("--lazy", action="store_true",
         help="switch to a lazy computation mode")
     parser.add_argument("--profiling", action="store_true",
         help="turn on detailed performance profiling")
+    parser.add_argument("--leap", action="store_true",
+        help="use leap timestepper")
     parser.add_argument("--overintegration", action="store_true",
         help="use overintegration in the RHS computations"),
     parser.add_argument("--esdg", action="store_true",
         help="use flux-differencing/entropy stable DG for inviscid computations.")
-    parser.add_argument("--leap", action="store_true",
-        help="use leap timestepper")
     parser.add_argument("--numpy", action="store_true",
         help="use numpy-based eager actx.")
     parser.add_argument("--restart_file", help="root name of restart file")
     parser.add_argument("--casename", help="casename to use for i/o")
     args = parser.parse_args()
-    from warnings import warn
-    warn("Automatically turning off DV logging. MIRGE-Com Issue(578)")
-    log_dependent = False
 
+    from warnings import warn
     from mirgecom.simutil import ApplicationOptionsError
     if args.esdg:
         if not args.lazy and not args.numpy:
@@ -466,13 +419,11 @@ if __name__ == "__main__":
     if args.casename:
         casename = args.casename
     rst_filename = None
-
     if args.restart_file:
         rst_filename = args.restart_file
 
     main(actx_class, use_leap=args.leap,
          casename=casename, rst_filename=rst_filename,
-         use_overintegration=args.overintegration or args.esdg,
-         use_esdg=args.esdg, log_dependent=log_dependent)
+         use_overintegration=args.overintegration or args.esdg, use_esdg=args.esdg)
 
 # vim: foldmethod=marker
