--- examples/wave-lazy.py
@@ -1,1 +1,0 @@
-wave.py
