--- examples/wave-mpi-lazy.py
@@ -1,1 +1,0 @@
-wave-mpi.py
